Update broken link to zplugin repo
The former link `https://github.com/psprint/zplugin.git` did not work. It came from the original repository, but that no longer exists as it once was.
That is, the link now points to my fork, since I don`t want to install other plugin managers, which are often quite bloated.